extract hand eigen model from utils
