+++ src/cpp/modules/manualEigen/HandEigenData.h
@@ -1,6 +1,6 @@
 #pragma once 
 
-#include "../../shared/utils.h"
+#include "../../shared/hand_eigen_model.h"
 
 #include <Eigen/Dense>
 #include <Eigen/StdVector>
