+++ src/cpp/modules/manualEigen/hand_eigen_d.cpp
@@ -10,6 +10,7 @@
 #include "hand_eigen_d.h"
 #include "../../shared/defs.h"
 #include "../../shared/utils.h"
+#include "../../shared/hand_eigen_model.h"
 
 using std::vector;
 using std::unordered_map;
